First pass at tables for sharing logs with the teams
This introduces the concept of a Session, which may or not map
exactly to an actual session of matches, that a given collection
of archives is used for. Running a Session then generates various
Artefacts which we want to share with the teams (logs, animations,
possibly other things). We assume that for some types of Artefact,
teams should only able to view those which relate to their robot.

With thanks to James for design discussion.

Co-Authored-By: James Seden Smith <sedders123@gmail.com>